Investigating the Relationship Between Parameters of EMG Signals and Muscular Force Utilizing Low-Pass Filtering
A biomedical signal known as the electromyography (EMG) signal detects electrical currents that the muscles produce while contracting, which are indicative of neuromuscular activities (Reaz, 2006). The EMG signal is a complex signal that depends on the physiological and anatomical characteristics of muscles and is regulated by the nervous system. The EMG signal picks up noise as it passes through various tissues, which could result in the interplay of various signals. Currently, the ability to detect EMG signals using effective and sophisticated techniques is becoming increasingly crucial in the field of biomedical engineering. 

This project utilizes various signal processing techniques and features in order to analyze the EMG signal. According to Zumbahlen (2008), the Butterworth filter is a class of filters that offers the optimum balance between phase response and attenuation. Parameters computed from the EMG signal include root mean square, which is typically used to evaluate the signal energy and amplitude in time domain (Patil et al., 2022), and fractal dimension of a time series, which characterizes the irregularity or randomness in the series (Breslin & Belward, 1999), along with mean and median frequency.